# auth_service/docs/swagger.yaml
basePath: /
...
host: localhost:8082
info:
  contact: {}
  description: Сервис авторизации
  title: Auth Service API
  version: "1.0"
paths:
...
  /auth/password/reset:
    post:
      consumes:
      - application/json
      description: |-
        Сбрасывает пароль пользователя с использованием токена,
        полученного по электронной почте.
        Токен должен быть передан в теле запроса в формате
        "selector.verifier", где selector — UUID, а verifier —
        строка в формате URL-safe Base64.
        Токен можно использовать только один раз. После успешного
        сброса пароля он становится недействительным.
        Новый пароль должен содержать не менее 8 символов и
        отличаться от текущего пароля.
      parameters:
      - description: Токен для сброса пароля и новый пароль
        in: body
        name: request
        required: true
        schema:
          properties:
            password:
              type: string
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Пароль успешно сброшен
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Некорректный формат токена, токен недействителен, истёк или
            уже был использован, пароль не соответствует требованиям либо совпадает
            с текущим
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      summary: Сброс пароля
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: |-
        ## Описание
        Обменивает действующий refresh токен на новую пару токенов (access и refresh).

        ### Процесс обновления:
        1. Валидация refresh токена (подпись JWT, срок действия)
        2. Извлечение user_id и app_id из payload токена
        3. Проверка что токен не находится в blacklist (Redis)
        4. Проверка статуса пользователя (не заблокирован, email подтвержден)
        5. Генерация новой пары токенов:
        - Новый access токен (TTL: 15 минут)
        - Новый refresh токен (TTL: 30 дней)
        6. Инвалидация старого refresh токена (token rotation)
        7. Сохранение нового refresh токена в БД

        ### Token Rotation (безопасность):
        - **Каждый refresh токен одноразовый** — после использования старый токен инвалидируется
        - Это защищает от атак с украденными токенами
        - При попытке использовать старый токен — все сессии пользователя инвалидируются

        ### Время жизни токенов:
        - **Access Token**: 15 минут (короткий для безопасности)
        - **Refresh Token**: 30 дней (удобство для пользователя)
        - Если пользователь неактивен 30 дней — требуется повторный login

        ### Когда использовать:
        - Access токен истек (получили 401 на защищенном endpoint)
        - Превентивное обновление перед истечением access токена
        - После длительного простоя приложения

        ### Ошибки:
        - `400`: Невалидный JSON или отсутствует refresh_token
        - `401`: Токен истек, невалиден или уже использован
        - `403`: Пользователь заблокирован
        - `500`: Ошибка БД или генерации токенов
      parameters:
      - description: Текущий refresh токен
        in: body
        name: token
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Новая пара токенов
          schema:
            properties:
              access_token:
                type: string
              refresh_token:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Невалидный или истекший токен
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Внутренняя ошибка
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      summary: Обновление access токена
      tags:
      - auth
      x-order: 3
  /auth/register:
    post:
      consumes:
      - application/json
      description: |-
        ## Описание
        Создает новую учетную запись пользователя в системе и отправляет письмо с подтверждением email.

        ### Процесс регистрации:
        1. Валидация входных данных (email формат, наличие username и пароля)
        2. Проверка уникальности email и username в базе данных
        3. Хеширование пароля с использованием bcrypt (cost factor 12)
        4. Создание записи пользователя в БД со статусом `email_verified = false`
        5. Генерация JWT токена верификации (валиден 24 часа)
        6. Отправка email с ссылкой подтверждения через RabbitMQ

        ### Требования к данным:
        - **Email**: Валидный email формат (example@domain.com), должен быть уникальным
        - **Username**: Минимум 3 символа, только буквы, цифры и подчеркивание, должен быть уникальным
        - **Password**: Минимум 8 символов, рекомендуется использовать заглавные буквы, цифры и спецсимволы

        ### Email верификация:
        - Письмо отправляется асинхронно через RabbitMQ (не блокирует ответ)
        - Токен верификации действует 24 часа
        - До подтверждения email пользователь не может войти в систему
        - Неподтвержденные аккаунты автоматически удаляются через 7 дней

        ### Формат ссылки верификации:
        `http://domain.com/auth/verify?token=eyJhbGc...`
      parameters:
      - description: Данные нового пользователя
        in: body
        name: user
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно создан, письмо отправлено
          schema:
            properties:
              status:
                type: string
              user_id:
                type: integer
            type: object
        "400":
          description: 'Ошибка валидации: некорректный email, слишком короткий пароль
            или отсутствуют обязательные поля'
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "409":
          description: Пользователь с таким email или username уже существует
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
        "500":
          description: 'Внутренняя ошибка: проблемы с БД, RabbitMQ или email сервисом'
          schema:
            properties:
              error:
                type: string
              status:
                type: string
            type: object
      summary: Регистрация нового пользователя
      tags:
      - auth
      x-order: 2
...
swagger: "2.0"